OLMIS-2158: fix window with report is blocked after sync with server
browsers by default blocks popups that are not a result of user's direct action. The call to window.open was made inside a callback function after sync with server. This makes it asynchronous and browser blocked it. The workaround was: instantiate window outside the callback function and you reference the variable to change the target url of that window in callback function when requisition is synced.