OLMIS-4004 Change call to hasRight to include facilityId
When checking if the user has a certain right, it should be checked by programId and facilityId, not programCode. Requisition rights must be checked by both program and facility, since they are supervision type rights. Additionally, we should move away from using codes, which should only be used for an upload process and are only available in an expanded REST response, and use ids, which are available in a normal REST response.