OLMIS-4004 Redo how permissions are checked in viewing a requisition
Remove logic from controller (since controllers should be “thin”), and add logic to methods in a factory, which are named generically (canSubmit, canDelete—can this user do this action on this requisition). Logic to check permissions no longer uses authorizationService (which is synchronous), but now uses permissionService (which is asynchronous).